Decouple Metrics Service creation from MobileCore
