+++ core/src/test/java/org/aerogear/mobile/core/integration/MetricsServiceIntegrationTest.java
@@ -29,7 +29,7 @@ public class MetricsServiceIntegrationTest {
     @Before
     public void setUp() {
         MobileCore.init(RuntimeEnvironment.application);
-        metricsService = MobileCore.getInstance().getService(MetricsService.class);
+        metricsService = new MetricsService();
 
         error = null;
     }
